requirements.txt
numpy
pyyaml
torch
matplotlib
